simulador_banco/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
connexion==3.2.0
coverage==7.8.0
cryptography==44.0.2
cssselect2==0.8.0
distlib==0.3.9
distro==1.9.0
dj-database-url==2.3.0
Django==5.2
django-bootstrap5==25.1
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-environ==0.12.0
django-heroku==0.3.1
django-markdownify==0.9.5
django-oauth-toolkit==3.0.1
django-sslserver==0.22
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-yasg==1.21.10
ecdsa==0.19.1
Faker==37.1.0
filelock==3.18.0
flake8==7.2.0
Flask==3.1.0
Flask-Testing==0.8.1
Flask-WTF==1.2.2
fonttools==4.57.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
kombu==5.5.4
livereload==2.7.1
lxml==5.3.2
Markdown==3.8
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
netifaces==0.11.0
nltk==3.9.1
oauthlib==3.2.2
openai==1.82.0
openapi-client==1.1.7
packaging==24.2
paramiko==3.5.1
pillow==11.2.1
platformdirs==4.3.7
playwright==1.52.0
pluggy==1.5.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydyf==0.11.0
pyee==13.0.0
pyflakes==3.3.2
PyJWT==2.9.0
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyphen==0.17.2
pyproject-api==1.9.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
pytest-mock==3.14.0
pytest-randomly==3.16.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
python-nmap==0.7.1
pytz==2025.2
PyYAML==6.0.2
qrcode==8.1
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.0
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9.1
selenium==3.141.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
swagger_ui_bundle==1.1.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
tornado==6.4.2
tox==4.25.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.16
vine==5.1.0
virtualenv==20.30.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.9.0
WTForms==3.2.1
zopfli==0.2.3.post1
gunicorn>=20.1
psycopg2-binary>=2.9      # Solo si usás PostgreSQL
python-dotenv>=1.0        # Manejo de variables de entorno
django-crispy-forms>=2.0  # Mejora de formularios en templates
PyJWT
celery
python-telegram-bot==20.6
pyotp==2.9.0